test: split config entry integration coverage
